fix and refactor state infrastructure
The bonuses from state infrastructure were applied by a fairly complicated script which ran every month on every state, applying or removing a bunch of flags and modifiers and so the state supply bonus matched the state infrastructure level. A rebalancing update in 2020 had reduced the supply bonuses, but hadn't altered the script for removing lower levels of bonus - this meant that higher levels of infrastructure was often massively reducing the level of local supply.

I've completely disabled the script which applied the supply bonus effects, and replaced it with a much simpler building effect, which has the same effect but working correctly, and much easier to maintain.

I've also moved the resource bonus into a building effect, and edited the script which adds bonus resources from infrastructure and efficiency to only add bonus resources from efficiency. I haven't disabled the script that calculates the state resource bonus, because the result of it (resource_modifier_set_startup in Resource_Efficiency.txt in common/scripted_effects) is used in the farm efficiency